+++ editor/editor-plugin-block-menu/src/blockMenuPluginType.ts
@@ -27,6 +27,11 @@ export type BlockMenuPlugin = NextEditorPlugin<
 			OptionalPlugin<DecorationsPlugin>,
 		];
 		pluginConfiguration?: BlockMenuPluginOptions;
+		sharedState:
+			| {
+					isFormatMenuHidden?: boolean;
+			  }
+			| undefined;
 	}
 >;