+++ editor/editor-plugin-block-menu/src/pm-plugins/main.ts
@@ -1,24 +1,30 @@
 import { SafePlugin } from '@atlaskit/editor-common/safe-plugin';
 import { PluginKey } from '@atlaskit/editor-prosemirror/state';
 
+import { getIsFormatMenuHidden } from './getIsFormatMenuHidden';
+
 export const blockMenuPluginKey = new PluginKey('blockMenuPlugin');
 
 // eslint-disable-next-line @typescript-eslint/no-empty-object-type
-type BlockMenuPluginState = {};
+type BlockMenuPluginState = {
+	isFormatMenuHidden: boolean;
+};
 
 export const createPlugin = () => {
 	return new SafePlugin<BlockMenuPluginState>({
 		key: blockMenuPluginKey,
 		state: {
 			init() {
-				return {};
+				return { isFormatMenuHidden: false };
 			},
 			apply: (tr, currentPluginState) => {
 				const meta = tr.getMeta(blockMenuPluginKey);
 				if (meta) {
-					return meta;
+					return { ...currentPluginState, ...meta };
 				}
-				return currentPluginState;
+				const isFormatMenuHidden = getIsFormatMenuHidden(tr);
+
+				return { ...currentPluginState, isFormatMenuHidden };
 			},
 		},
 	});