+++ editor/editor-plugin-block-menu/src/ui/format-menu-section.tsx
@@ -1,0 +1,23 @@
+import React from 'react';
+
+import type { ExtractInjectionAPI } from '@atlaskit/editor-common/types';
+import { useSharedPluginStateSelector } from '@atlaskit/editor-common/use-shared-plugin-state-selector';
+import { ToolbarDropdownItemSection } from '@atlaskit/editor-toolbar';
+
+import type { BlockMenuPlugin } from '../blockMenuPluginType';
+
+export const FormatMenuSection = ({
+	children,
+	api,
+}: {
+	api: ExtractInjectionAPI<BlockMenuPlugin> | undefined;
+	children: React.ReactNode;
+}) => {
+	const isFormatMenuHidden = useSharedPluginStateSelector(api, 'blockMenu.isFormatMenuHidden');
+
+	if (isFormatMenuHidden) {
+		return null;
+	}
+
+	return <ToolbarDropdownItemSection>{children}</ToolbarDropdownItemSection>;
+};