+++ linking-platform/link-datasource/src/ui/confluence-search-modal/modal/messages.ts
@@ -57,4 +57,11 @@ export const confluenceSearchModalMessages = defineMessages({
 		defaultMessage:
 			'Check your connection and refresh, or <a>open this project in Confluence</a> to review.',
 	},
+	checkConnectionWithSourceGalaxia: {
+		id: 'linkDataSource.confluence-search.checkConnectionWithSource-galaxia',
+		description:
+			'Instructions to let the user know how to resolve the error that occured, or click the link provided to open the query in Jira',
+		defaultMessage:
+			'Check your connection and refresh, or <a>open this Jira space in Confluence</a> to review.',
+	},
 });