+++ editor/editor-plugins/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-plugins
 
+## 10.3.16
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 10.3.15
 
 ### Patch Changes